api/requirements.txt
fastapi
uvicorn
pandas
psycopg2
pyjwt
requests
SQLAlchemy
